# openspec/contracts/content-workflows/openapi/v1.yaml
openapi: 3.1.0
info:
  title: ACA Canonical Workflow API
  version: 2.0.0
  description: |
    Coordinated breaking replacement for ingestion, summarization, theme,
    digest, pipeline, podcast, and audio mutation contracts. All mutations
    return a durable OperationHandle backed by pgqueuer_jobs.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.aca.rotkohl.ai
    description: Production
security:
  - AdminKey: []
paths:
  /api/v1/capabilities:
    get:
      operationId: getCapabilities
      summary: Discover source, operation, resource, and transport capabilities
      tags: [capabilities]
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200":
          description: Deployment capabilities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CapabilityDocument" }
              example:
                contract_version: "2.0.0"
                source_commands:
                  - key: rss
                    display_name: RSS
                    emitted_sources: [rss]
                    scheduled: true
                    transports: [cli, http, mcp, frontend]
                    fields:
                      - { name: max_items, type: integer, required: false }
                operation_types: [ingestion.execute, digest.create]
                resource_types: [content, digest]
                next_cursor: null
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/configured-sources:
    get:
      operationId: listConfiguredSources
      summary: List configured ingestion sources without assuming locator fields
      tags: [capabilities, ingestion]
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200":
          description: Configured source page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfiguredSourcePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/v1/operations/{operation_id}:
    get:
      operationId: getOperation
      summary: Get operation state, optionally waiting for change or completion
      tags: [operations]
      parameters:
        - $ref: "#/components/parameters/OperationId"
        - name: wait_seconds
          in: query
          required: false
          schema: { type: integer, minimum: 0, maximum: 30, default: 0 }
          example: 10
      responses:
        "200":
          description: Current operation state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationHandle" }
              example:
                schema_version: 2
                operation_id: "8123"
                operation_type: digest.create
                status: completed
                progress: 100
                message: Digest created
                cancellable: false
                retry_count: 0
                status_url: /api/v1/operations/8123
                events_url: /api/v1/operations/8123/events
                resource: { type: digest, id: "42", url: /api/v1/digests/42 }
                result: { selection_fingerprint: "b6d81b360a5672d80c27430f39153e2c" }
                problem: null
                created_at: "2026-07-13T10:00:00Z"
                started_at: "2026-07-13T10:00:01Z"
                completed_at: "2026-07-13T10:01:12Z"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFoundProblem" }
  /api/v1/operations:
    get:
      operationId: listOperations
      summary: List bounded operation summaries using an opaque keyset cursor
      tags: [operations]
      parameters:
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/OperationStatus" }
          description: Optional lifecycle filter used for bounded active hydration.
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200":
          description: Operation page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OperationPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationProblem" }
  /api/v1/workflow-alert-verification-context:
    get:
      operationId: getWorkflowAlertVerificationContext
      summary: Read positive staging deployment identity for destructive alert verification
      tags: [operations, observability]
      responses:
        "200":
          description: Authenticated staging deployment identity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkflowAlertVerificationContext" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ProblemResponse" }
  /api/v1/workflow-terminal-events/{event_id}:
    get:
      operationId: getWorkflowTerminalEvent
      summary: Read one bounded sanitized workflow terminal-event diagnostic
      tags: [operations, observability]
      parameters:
        - name: event_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Sanitized terminal-event diagnostic
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkflowTerminalEventDiagnostic" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFoundProblem" }
        "422": { $ref: "#/components/responses/ValidationProblem" }
  /api/v1/operations/reconcile-content:
    post:
      operationId: reconcileContent
      summary: Preview or apply one bounded page of content reconciliation
      tags: [operations]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContentReconciliationRequest" }
            example: { apply: false, limit: 50, after_content_id: null }
      responses:
        "200":
          description: Bounded content reconciliation report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContentReconciliationReport" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationProblem" }
        "503":
          description: Reconciliation apply is disabled by server policy
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/operations/{operation_id}/events:
    get:
      operationId: streamOperationEvents
      summary: Stream operation progress events
      tags: [operations]
      parameters:
        - $ref: "#/components/parameters/OperationId"
        - name: Last-Event-ID
          in: header
          required: false
          schema: { type: string }
          example: "8123:4"
      responses:
        "200":
          description: Server-sent operation events
          content:
            text/event-stream:
              schema: { $ref: "#/components/schemas/OperationEvent" }
              example: "id: 8123:5\nevent: progress\ndata: {\"operation_id\":\"8123\",\"status\":\"in_progress\",\"progress\":65,\"message\":\"Creating digest\"}\n\n"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFoundProblem" }
  /api/v1/operations/{operation_id}/retry:
    post:
      operationId: retryOperation
      summary: Retry a failed operation
      tags: [operations]
      parameters:
        - $ref: "#/components/parameters/OperationId"
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
            example: {}
      responses:
        "202": { $ref: "#/components/responses/AcceptedOperation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFoundProblem" }
        "409": { $ref: "#/components/responses/ConflictProblem" }
  /api/v1/operations/{operation_id}/cancel:
    post:
      operationId: cancelOperation
      summary: Cancel a queued operation or request safe running cancellation
      tags: [operations]
      parameters:
        - $ref: "#/components/parameters/OperationId"
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
            example: {}
      responses:
        "202": { $ref: "#/components/responses/AcceptedOperation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFoundProblem" }
        "409": { $ref: "#/components/responses/ConflictProblem" }
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  parameters:
    PageLimit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    PageCursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
...
    OperationId:
      name: operation_id
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9]+$" }
      example: "8123"
  responses:
    AcceptedOperation:
      description: Durable operation accepted or existing idempotent operation returned
      content:
        application/json:
          schema: { $ref: "#/components/schemas/OperationHandle" }
          example:
            schema_version: 2
            operation_id: "8123"
            operation_type: digest.create
            status: queued
            progress: 0
            message: Queued
            cancellable: true
            retry_count: 0
            status_url: /api/v1/operations/8123
            events_url: /api/v1/operations/8123/events
            resource: null
            result: null
            problem: null
            created_at: "2026-07-13T10:00:00Z"
            started_at: null
            completed_at: null
...
    ProblemResponse:
      description: RFC 7807 problem
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://aca.rotkohl.ai/problems/request-failed
            title: Request failed
            status: 400
            detail: The request could not be completed
    Unauthorized:
      description: Authentication is required
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://aca.rotkohl.ai/problems/unauthorized
            title: Unauthorized
            status: 401
            detail: Valid credentials are required
    Forbidden:
      description: Credentials do not authorize the operation
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://aca.rotkohl.ai/problems/forbidden
            title: Forbidden
            status: 403
            detail: The supplied credentials are not authorized
    ValidationProblem:
      description: Request validation failed
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://aca.rotkohl.ai/problems/validation
            title: Validation failed
            status: 422
            detail: arxiv_paper.identifier is required
            errors:
              - { path: [identifier], code: missing, message: Field required }
    ConflictProblem:
      description: Current resource or operation state conflicts with the request
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://aca.rotkohl.ai/problems/operation-conflict
            title: Operation conflict
            status: 409
            detail: Completed operations cannot be cancelled
    NotFoundProblem:
      description: Requested operation or resource was not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
          example:
            type: https://aca.rotkohl.ai/problems/not-found
            title: Not found
            status: 404
            detail: Operation 8123 was not found
  schemas:
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: [string, "null"] }
        code: { type: [string, "null"] }
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [path, code, message]
            properties:
              path: { type: array, items: { type: [string, integer] } }
              code: { type: string }
              message: { type: string }
    ResourceReference:
      type: object
      additionalProperties: false
      required: [type, id, url]
      properties:
        type:
          type: string
          enum: [content, ingestion_run, summary_batch, theme_analysis, digest, pipeline_run, podcast_script, podcast, audio_digest]
        id: { type: string }
        url: { type: string }
...
    OperationStatus:
      type: string
      enum: [queued, in_progress, completed, failed, cancelled]
...
    IngestionResult:
      oneOf:
        - { $ref: "#/components/schemas/IngestionResultV1" }
        - { $ref: "#/components/schemas/IngestionResultV2" }
      description: >-
        Legacy V1 rows may predate an explicit schema_version, so readers use
        structural union validation rather than a discriminator.
...
    PipelineResultV2:
      type: object
      additionalProperties: true
      x-operation-type: pipeline.run
      required: [schema_version, ingestion_summary]
      properties:
        schema_version: { type: integer, const: 2 }
        ingestion_summary: { $ref: "#/components/schemas/PipelineIngestionSummary" }
...
    OperationSummary:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - operation_id
        - operation_type
        - status
        - progress
        - message
        - cancellable
        - retry_count
        - status_url
        - events_url
        - created_at
      properties:
        schema_version: { type: integer, const: 2 }
        operation_id: { type: string, pattern: "^[1-9][0-9]*$", maxLength: 19 }
        operation_type:
          type: string
          enum:
            - ingestion.execute
            - summarization.run
            - theme_analysis.create
            - digest.create
            - pipeline.run
            - podcast_script.create
            - podcast_audio.create
            - audio_digest.create
        status: { $ref: "#/components/schemas/OperationStatus" }
        progress: { type: integer, minimum: 0, maximum: 100 }
        message: { type: string, maxLength: 500 }
        cancellable: { type: boolean }
        retry_count: { type: integer, minimum: 0 }
        status_url: { type: string, pattern: "^/api/v1/operations/[1-9][0-9]*$" }
        events_url: { type: string, pattern: "^/api/v1/operations/[1-9][0-9]*/events$" }
        created_at: { type: string, format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        completed_at: { type: [string, "null"], format: date-time }
    OperationHandle:
      type: object
      additionalProperties: false
      x-operation-result-schemas:
        ingestion.execute: { $ref: "#/components/schemas/IngestionResult" }
        pipeline.run: { $ref: "#/components/schemas/PipelineResultV2" }
      required:
        - schema_version
        - operation_id
        - operation_type
        - status
        - progress
        - message
        - cancellable
        - retry_count
        - status_url
        - events_url
        - created_at
      properties:
        schema_version: { type: integer, const: 2 }
        operation_id: { type: string }
        operation_type:
          type: string
          enum:
            - ingestion.execute
            - summarization.run
            - theme_analysis.create
            - digest.create
            - pipeline.run
            - podcast_script.create
            - podcast_audio.create
            - audio_digest.create
        status: { $ref: "#/components/schemas/OperationStatus" }
        progress: { type: integer, minimum: 0, maximum: 100 }
        message: { type: string }
        cancellable: { type: boolean }
        retry_count: { type: integer, minimum: 0 }
        status_url: { type: string }
        events_url: { type: string }
        resource:
          anyOf:
            - { $ref: "#/components/schemas/ResourceReference" }
            - { type: "null" }
        result:
          type: [object, "null"]
          additionalProperties: true
        problem:
          anyOf:
            - { $ref: "#/components/schemas/Problem" }
            - { type: "null" }
        created_at: { type: string, format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        completed_at: { type: [string, "null"], format: date-time }
    OperationPage:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/OperationSummary" }
        next_cursor: { type: [string, "null"], maxLength: 2048 }
    WorkflowAlertVerificationContext:
      type: object
      additionalProperties: false
      required: [schema_version, environment_class, revision, revision_source]
      properties:
        schema_version: { type: integer, const: 1 }
        environment_class: { type: string, const: staging }
        revision: { type: string, minLength: 40, maxLength: 40, pattern: "^[a-f0-9]{40}$" }
        revision_source: { type: string, const: railway_commit_sha }
    WorkflowTerminalDeliveryCounts:
      type: object
      additionalProperties: false
      required: [pending, leased, delivered, permanent_failure, exhausted]
      properties:
        pending: { type: integer, minimum: 0, maximum: 9223372036854775807 }
        leased: { type: integer, minimum: 0, maximum: 9223372036854775807 }
        delivered: { type: integer, minimum: 0, maximum: 9223372036854775807 }
        permanent_failure: { type: integer, minimum: 0, maximum: 9223372036854775807 }
        exhausted: { type: integer, minimum: 0, maximum: 9223372036854775807 }
    WorkflowTerminalEventDiagnostic:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - event_id
        - event_key
        - source_kind
        - operation_id
        - claim_generation
        - terminal_status
        - classification_status
        - release_revision
        - release_revision_source
        - occurred_at
        - telemetry_emitted_at
        - delivery_counts
      properties:
        schema_version: { type: integer, const: 1 }
        event_id: { type: string, format: uuid }
        event_key: { type: string, minLength: 1, maxLength: 160, pattern: "^[a-z0-9:_-]+$" }
        source_kind:
          type: string
          enum: [operation, reconciliation_action, reconciliation_failure]
        operation_id:
          type: [string, "null"]
          maxLength: 19
          pattern: "^[1-9][0-9]*$"
        claim_generation:
          type: [integer, "null"]
          minimum: 0
          maximum: 2147483647
        terminal_status:
          type: [string, "null"]
          enum: [completed, failed, cancelled, null]
        classification_status:
          type: string
          enum: [pending, ready, telemetry_only, rejected]
        release_revision:
          type: [string, "null"]
          maxLength: 40
          pattern: "^(?:[a-f0-9]{40}|development|unavailable)$"
        release_revision_source:
          type: [string, "null"]
          enum: [railway_commit_sha, local_development, unavailable, null]
        occurred_at: { type: string, format: date-time }
        telemetry_emitted_at: { type: [string, "null"], format: date-time }
        delivery_counts: { $ref: "#/components/schemas/WorkflowTerminalDeliveryCounts" }
    ContentReconciliationMode:
      type: string
      enum: [dry_run, apply]
...
    ContentReconciliationRequest:
      type: object
      additionalProperties: false
      properties:
        apply: { type: boolean, default: false }
        limit: { type: integer, minimum: 1, maximum: 100 }
        after_content_id:
          type: [integer, "null"]
          minimum: 1
          maximum: 9223372036854775807
...
    ContentReconciliationReport:
      type: object
      additionalProperties: false
      required: [run_id, mode, scanned, reported, counts, items]
      properties:
        run_id: { type: string, format: uuid }
        mode: { $ref: "#/components/schemas/ContentReconciliationMode" }
        scanned: { type: integer, minimum: 0, maximum: 100 }
        reported: { type: integer, minimum: 0, maximum: 100 }
        next_after_content_id:
          type: [integer, "null"]
          minimum: 1
          maximum: 9223372036854775807
        counts: { $ref: "#/components/schemas/ContentReconciliationCounts" }
        items:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/ContentReconciliationItem" }
    OperationEvent:
      type: object
      additionalProperties: false
      required: [schema_version, event_id, operation_id, operation_type, status, progress, message, occurred_at]
      properties:
        schema_version: { type: integer, const: 2 }
        event_id: { type: string }
        operation_id: { type: string }
        operation_type:
          type: string
          enum:
            - ingestion.execute
            - summarization.run
            - theme_analysis.create
            - digest.create
            - pipeline.run
            - podcast_script.create
            - podcast_audio.create
            - audio_digest.create
        status: { $ref: "#/components/schemas/OperationStatus" }
        progress: { type: integer, minimum: 0, maximum: 100 }
        message: { type: string }
        resource:
          anyOf:
            - { $ref: "#/components/schemas/ResourceReference" }
            - { type: "null" }
        problem:
          anyOf:
            - { $ref: "#/components/schemas/Problem" }
            - { type: "null" }
        occurred_at: { type: string, format: date-time }
...
    CapabilityField:
      type: object
      additionalProperties: false
      required: [name, type, required]
      properties:
        name: { type: string }
        type: { type: string }
        required: { type: boolean }
        description: { type: [string, "null"] }
        format: { type: [string, "null"] }
        enum: { type: array, items: { type: string } }
        default: {}
        constraints:
          type: object
          additionalProperties: true
          default: {}
    SourceCapability:
      type: object
      additionalProperties: false
      required:
        - key
        - display_name
        - emitted_sources
        - scheduled
        - transports
        - supports_force
        - supports_date_range
        - supports_preview
        - requires_identifier
        - fields
      properties:
        key: { type: string }
        display_name: { type: string }
        emitted_sources:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string }
        scheduled: { type: boolean }
        supports_force: { type: boolean }
        supports_date_range: { type: boolean }
        supports_preview: { type: boolean }
        requires_identifier: { type: boolean }
        transports:
          type: array
          items: { type: string, enum: [cli, http, mcp, frontend] }
          uniqueItems: true
        fields: { type: array, items: { $ref: "#/components/schemas/CapabilityField" } }
    CapabilityDocument:
      type: object
      additionalProperties: false
      required: [contract_version, source_commands, operation_types, resource_types]
      properties:
        contract_version: { type: string }
        source_commands: { type: array, items: { $ref: "#/components/schemas/SourceCapability" } }
        operation_types: { type: array, items: { type: string } }
        resource_types: { type: array, items: { type: string } }
        next_cursor: { type: [string, "null"] }
    ConfiguredSource:
      type: object
      additionalProperties: false
      required: [key, command_key, source_type, enabled, origin, configuration]
      properties:
        key: { type: string }
        command_key: { type: string }
        source_type: { type: string }
        name: { type: [string, "null"] }
        enabled: { type: boolean }
        origin: { type: string, enum: [yaml, db] }
        configuration: { type: object, additionalProperties: true }
    ConfiguredSourcePage:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/ConfiguredSource" } }
        next_cursor: { type: [string, "null"] }
...